Made sleuther framework attempt to resume webhooks on startup.
Also refactored tests.
