+++ magda-sleuther-framework/src/getWebhookUrl.ts
@@ -1,0 +1,5 @@
+import SleutherOptions from "./SleutherOptions";
+
+export default function getHookUrl(options: SleutherOptions) {
+    return `${options.argv.internalUrl}/hook`;
+}
