+++ magda-sleuther-framework/src/test/fakeArgv.ts
@@ -1,0 +1,15 @@
+import { SleutherArguments } from "../commonYargs";
+
+export default function fakeArgv(options: {
+    internalUrl: string;
+    registryUrl: string;
+    jwtSecret: string;
+    userId: string;
+    listenPort: number;
+}): SleutherArguments {
+    return {
+        ...options,
+        $0: "",
+        _: []
+    };
+}
